requirements.txt
numpy==1.24.4
tqdm==4.67.1
ipywidgets>=7.0,<8.0
ipykernel==6.29.2
seaborn==0.13.2
matplotlib==3.7.1
hydra-core==1.3.2 

transformers==4.51.3 
accelerate==1.6.0 
datasets==2.16.1
bitsandbytes==0.45.1
wandb==0.19.4 
evaluate==0.4.3 
nltk==3.9.1 
rouge-score==0.1.2 
absl-py==2.1.0 
peft==0.9.0 
python-dotenv==1.1.0 
deepspeed==0.13.1
conbo==0.1.0

# torch
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.3.1
torchvision 
torchaudio 